Write gscramble I_meta, M_meta, and Geno to a plink file

Writes genetic and individual information in gscramble's I_meta, M_meta, and Geno like objects into a uncompressed plink .ped and .map files.

gscramble2plink(I_meta, M_meta, Geno, prefix = tempfile())

Arguments

I_meta

a tibble of individual meta data with at least the columns of group and indiv.

M_meta

a tibble of marker meta data with at least the columns of chrom, pos, and variant_id.

Geno

a character matrix of genotypes. Num-indivs rows and num-markers * 2 columns, with missing denoted denoted by NA.

prefix

the file path and prefix into which to write out the files

Value

Returns TRUE if successful.

Examples

gscramble2plink(I_meta, M_meta, Geno)
#> pedfile written to /var/folders/xg/mz_qt7q54yv_hwzvhskwx2c00000gp/T//Rtmp7jZJ1q/filefcf07962108d.ped
#> mapfile written to /var/folders/xg/mz_qt7q54yv_hwzvhskwx2c00000gp/T//Rtmp7jZJ1q/filefcf07962108d.map
#> [1] TRUE