Take the output of rearrange_genos and permute everyone by population

This is done prior to assign random genomic fragments of individuals in the sample to the founders of the GSP, to be dropped to the samples.

perm_gs_by_pops(GS, preserve_haplotypes = FALSE)

Arguments

GS

the tibble that is the output from rearrange_genos

preserve_haplotypes

If true then the Geno data is assumed phased (first allele at an individual on one haplotype and second allele on the other) and those haplotypes are preserved in this permutation of genomic material amongst the founders.

Examples

# first get the output of rearrange_genos
RG <- rearrange_genos(Geno, I_meta, M_meta)

# then permute by the populations
PG <- perm_gs_by_pops(RG)