computes the simulation-level founder haplotype index for each founder haplo

This takes the output of segregate() and deals with the multiple gpp's and reps to come up with a unique index for each found haplotype, so that those haplotypes can all, eventually, be accessed easily out of the genotype matrix. Along the way, this function does some light checking to make sure that the rs_founder_haplo values are dense within gpp and rep as they should be.

Usage

sim_level_founder_haplos(S)

Arguments

S

tibble of segments like that produced by segregate.

Value

This function returns a result that is basically the output of segregate() with an additional column added to it: sim_level_founder_haplo. This is the index of the haplotype within each group_origin that should be used. For details of the other columns in the output tibble, see the documentation for segregate.

Examples

#### Get output from segregate to use as input ####
# We construct an example here where we will request segregation
# down a GSP with two F1s and F1B backcrosses between two hypothetical
# populations, A and B.
gsp_f1f1b <- create_GSP("A", "B", F1 = TRUE, F1B = TRUE)

# We will imagine that in our marker data there are three groups
# labelled "grp1", "grp2", and "grp3", and we want to create the F1Bs with backcrossing
# only to grp3.
reppop <- tibble::tibble(
  rep = c(1, 1, 2, 2),
  pop = c("A", "B", "A", "B"),
  group = c("grp3", "grp1", "grp3", "grp2")
)

# combine those into a request
request <- tibble::tibble(
  gpp = list(gsp_f1f1b),
  reppop = list(reppop)
)

# now run it through segregate()
set.seed(5)  # just for reproducibility in example...
simSegs <- segregate(request, RecRates)

#### Now we can run those through sim_level_founder_haplos() ####
fh <- sim_level_founder_haplos(simSegs)
fh
#> # A tibble: 52 × 14
#>      gpp   rep chrom ped_sample_id samp_index gamete_index gamete_segments
#>    <int> <dbl> <chr> <chr>              <int>        <dbl> <list>         
#>  1     1     1 12    4                      1            1 <dbl [2]>      
#>  2     1     1 17    4                      1            1 <dbl [2]>      
#>  3     1     1 18    4                      1            1 <dbl [2]>      
#>  4     1     1 12    4                      1            2 <dbl [2]>      
#>  5     1     1 17    4                      1            2 <dbl [2]>      
#>  6     1     1 18    4                      1            2 <dbl [2]>      
#>  7     1     1 12    5                      1            1 <dbl [2]>      
#>  8     1     1 17    5                      1            1 <dbl [2]>      
#>  9     1     1 18    5                      1            1 <dbl [2]>      
#> 10     1     1 12    5                      1            2 <dbl [3]>      
#> # … with 42 more rows, and 7 more variables: pop_origin <chr>,
#> #   rs_founder_haplo <int>, start <dbl>, end <dbl>, group_origin <chr>,
#> #   sim_level_founder_haplo.x <int>, sim_level_founder_haplo.y <int>